Add support for recursive wildcard pattern in .dockerignore
