+++ docker/utils/__init__.py
@@ -1,7 +1,9 @@
 # flake8: noqa
+from .build import tar, exclude_paths
+from .decorators import check_resource, minimum_version, update_headers
 from .utils import (
     compare_version, convert_port_bindings, convert_volume_binds,
-    mkbuildcontext, tar, exclude_paths, parse_repository_tag, parse_host,
+    mkbuildcontext, parse_repository_tag, parse_host,
     kwargs_from_env, convert_filters, datetime_to_timestamp,
     create_host_config, parse_bytes, ping_registry, parse_env_file, version_lt,
     version_gte, decode_json_header, split_command, create_ipam_config,
@@ -9,4 +11,3 @@ from .utils import (
     format_environment, create_archive
 )
 
-from .decorators import check_resource, minimum_version, update_headers
